requirements.txt
asgiref==3.5.2
boto3==1.24.20
botocore==1.27.20
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-storages==1.12.3
Django-Verify-Email==2.0.3
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe==3.2.0
tzdata==2022.1
urllib3==1.26.9
python-dotenv